src/requirements.txt
# asyncpg==0.22.0
fastapi==0.68.1
psycopg2-binary==2.8.6
python-dotenv==0.19.0
python-multipart==0.0.5
boto3==1.16.13
SQLAlchemy==1.4.23
pymysql==1.0.2
google-auth==1.26.1

PyJWT==1.6.1
uvicorn==0.15.0

# dev
pytest==6.2.3
requests==2.25.1







